Improve tile compressed pixel access API
BinaryTableData is an enum that contains 2 variants:
* The TableData traditional API
* The TileCompressed variant containing Pixels data. Pixels is an enum that can contain U8, I16, I32, F32 uncompressed pixels. You get an iterator over the uncompressed native u8, i16, i32, f32 pixels.
