+++ src/hdu/data/bintable/tile_compressed/dithering.rs
